properly handle embedded structs from another package
If an embedded struct comes from another package, and it has fields
referring to identifiers in that other package, this change makes the
generated code correctly refer to those identifiers using the name of
the imported package.

This turns out to be needed for go-algorand PR #1856, since the agreement
package has an embedded field of type bookkeeping.BlockHeader, which in
turn now has a field of type map[...]CompactCertState.  The generated msgp
code needs to refer to the map elements as bookkeeping.CompactCertState.